Resolved conversation

+++ src/main/java/org/prebid/server/proto/openrtb/ext/request/gumgum/ExtImpGumgum.java
@@ -18,5 +18,5 @@ public class ExtImpGumgum {
     @JsonProperty("irisid")
     String irisId;
 
-    BigInteger slot;
+    Long slot;
 }
