+++ src/main/java/org/prebid/server/proto/openrtb/ext/request/gumgum/ExtImpGumgumBanner.java
@@ -3,13 +3,11 @@ package org.prebid.server.proto.openrtb.ext.request.gumgum;
 import lombok.AllArgsConstructor;
 import lombok.Value;
 
-import java.math.BigInteger;
-
 @Value
 @AllArgsConstructor(staticName = "of")
 public class ExtImpGumgumBanner {
 
-    BigInteger slot;
+    Long slot;
 
     Integer maxw;
